requirements.txt
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.5
python-certifi-win32==1.6.1
coloredlogs==15.0.1
pandas==2.1.1
openpyxl==3.1.2
enlighten==1.11.2
beautifulsoup4==4.12.2
lxml==4.9.3
